SW-2341 Dead/Other withdrawal overview
- extracted out the overview item card from outplant
- combined dead and other into DeadOtherContent (no distinction for now)
- I can keep them separate if we want that separation
- if this approach is fine, I can follow up with a table PR
