--- src/components/NurseryWithdrawals/WithdrawalDetails/DeadContent.tsx
@@ -1,16 +1,0 @@
-import { Box, Typography } from '@mui/material';
-import strings from 'src/strings';
-import Photos from './sections/Photos';
-
-type DeadContentProps = {};
-
-export default function DeadContent(props: DeadContentProps): JSX.Element {
-  return (
-    <Box display='flex' flexDirection='column'>
-      <Typography fontSize='20px' fontWeight={600}>
-        {strings.DETAILS}
-      </Typography>
-      <Photos />
-    </Box>
-  );
-}
